fix collection form
